1번 회원 관리자로 지정 게시물 수정,삭제시 권한체크

+++ src/main/java/com/sbs/untact/controller/UsrArticleController.java
@@ -3,6 +3,8 @@ package com.sbs.untact.controller;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.sbs.untact.dto.Article;
 import com.sbs.untact.dto.ResultData;
 import com.sbs.untact.service.ArticleService;
+import com.sbs.untact.util.Util;
 
 @Controller
 public class UsrArticleController {
@@ -43,7 +46,11 @@ public class UsrArticleController {
 	}
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData doAdd(@RequestParam Map<String, Object> param) {
+	public ResultData doAdd(@RequestParam Map<String, Object> param, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		if(loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
 		if (param.get("title") == null) {
 			return new ResultData("F-1", "title을 입력해주세요.");
 		}
@@ -51,22 +58,36 @@ public class UsrArticleController {
 		if (param.get("title") == null) {
 			return new ResultData("F-1", "body를 입력해주세요.");
 		}
-		
+		param.put("memberId", loginedMemberId);
 		return articleService.addArticle(param);
 	}
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		if(loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
+		if(id == null) {
+			return new ResultData("F-2", "게시물 번호를 입력해주세요");
+		}
 		Article article = articleService.getArticle(id);
 		if(article == null) {
 			return new ResultData("F-1", "존재하지 않는 게시물입니다");
 		}
-		
+		ResultData actorCanDeleteRd = articleService.getActorCanDeleteRd(article, loginedMemberId);
+		if(actorCanDeleteRd.isFail()) {
+			return actorCanDeleteRd;
+		}
 		return articleService.deleteArticle(id);
 	}
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		if(loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인 후 이용해주세요.");
+		}
 		if(id == null) {
 			return new ResultData("F-1", "id를 입력해주세요");
 		}
@@ -75,6 +96,14 @@ public class UsrArticleController {
 		}
 		if(body == null) {
 			return new ResultData("F-1", "body를 입력해주세요");
+		}
+		Article article = articleService.getArticle(id);
+		if(article == null) {
+			return new ResultData("F-1", "존재하지 않는 게시물 입니다");
+		}
+		ResultData actorCanModifyRd = articleService.getActorCanModifyRd(article, loginedMemberId);
+		if(actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
 		}
 		return articleService.modifyArticle(id, title, body);
 	}
